Move `extensions/pkg/controller/worker/helper.BuildOwnerToMachine{Set}sMap` functions to `pkg/utils/gardener` package
They will be reused from other (to-be-introduced) functions in the `pkg/utils/gardener` packages, and we don't want to import the extensions library from there. Hence, let's better move them.
