+++ extensions/pkg/controller/worker/genericactuator/actuator_test.go
@@ -30,7 +30,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/gardener/gardener/extensions/pkg/controller/worker"
-	extensionsworkerhelper "github.com/gardener/gardener/extensions/pkg/controller/worker/helper"
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 )
 
@@ -128,7 +127,7 @@ var _ = Describe("Actuator", func() {
 	Describe("#isMachineControllerStuck", func() {
 		var (
 			machineDeploymentName           = "machine-deployment-1"
-			machineDeploymentOwnerReference = []metav1.OwnerReference{{Name: machineDeploymentName, Kind: extensionsworkerhelper.MachineDeploymentKind}}
+			machineDeploymentOwnerReference = []metav1.OwnerReference{{Name: machineDeploymentName, Kind: "MachineDeployment"}}
 
 			machineClassName      = "machine-class-new"
 			machineDeploymentSpec = machinev1alpha1.MachineDeploymentSpec{
