+++ pkg/utils/gardener/machines.go
@@ -1,0 +1,67 @@
+// Copyright 2023 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package gardener
+
+import (
+	machinev1alpha1 "github.com/gardener/machine-controller-manager/pkg/apis/machine/v1alpha1"
+)
+
+const (
+	nameLabel = "name"
+	// MachineSetKind is the kind of the owner reference of a machine set
+	MachineSetKind = "MachineSet"
+	// MachineDeploymentKind is the kind of the owner reference of a machine deployment
+	MachineDeploymentKind = "MachineDeployment"
+)
+
+// BuildOwnerToMachinesMap builds a map from a slice of machinev1alpha1.Machine, that maps the owner reference
+// to a slice of machines with the same owner reference
+func BuildOwnerToMachinesMap(machines []machinev1alpha1.Machine) map[string][]machinev1alpha1.Machine {
+	ownerToMachines := make(map[string][]machinev1alpha1.Machine)
+	for index, machine := range machines {
+		if len(machine.OwnerReferences) > 0 {
+			for _, reference := range machine.OwnerReferences {
+				if reference.Kind == MachineSetKind {
+					ownerToMachines[reference.Name] = append(ownerToMachines[reference.Name], machines[index])
+				}
+			}
+		} else if len(machine.Labels) > 0 {
+			if machineDeploymentName, ok := machine.Labels[nameLabel]; ok {
+				ownerToMachines[machineDeploymentName] = append(ownerToMachines[machineDeploymentName], machines[index])
+			}
+		}
+	}
+	return ownerToMachines
+}
+
+// BuildOwnerToMachineSetsMap builds a map from a slice of machinev1alpha1.MachineSet, that maps the owner reference
+// to a slice of MachineSets with the same owner reference
+func BuildOwnerToMachineSetsMap(machineSets []machinev1alpha1.MachineSet) map[string][]machinev1alpha1.MachineSet {
+	ownerToMachineSets := make(map[string][]machinev1alpha1.MachineSet)
+	for index, machineSet := range machineSets {
+		if len(machineSet.OwnerReferences) > 0 {
+			for _, reference := range machineSet.OwnerReferences {
+				if reference.Kind == MachineDeploymentKind {
+					ownerToMachineSets[reference.Name] = append(ownerToMachineSets[reference.Name], machineSets[index])
+				}
+			}
+		} else if len(machineSet.Labels) > 0 {
+			if machineDeploymentName, ok := machineSet.Labels[nameLabel]; ok {
+				ownerToMachineSets[machineDeploymentName] = append(ownerToMachineSets[machineDeploymentName], machineSets[index])
+			}
+		}
+	}
+	return ownerToMachineSets
+}
